fix(money): close the Float->Decimal gap end-to-end; wire Sprint 6 events
The Bug 2 precision fix (quotation service, prior commit) never actually
survived an HTTP hop: the Pydantic response schema still typed money
fields as float, and payments-data's Payment.amount column was still
Float outright. SRS section 6 requires payment amount to match the
quotation total exactly - that check was float-vs-float the whole time.

Money precision, end to end:
- quotation: QuotationResponse/LineItem money fields -> Decimal
- payments-data: Payment.amount column -> Numeric(12,2) (new migration
  0002), PaymentCreateRequest/PaymentResponse -> Decimal, fetch_quotation
  parses the quotation total as Decimal instead of trusting response.json()
- Discovered empirically: FastAPI serializes Decimal as a JSON *string*,
  not a number. Updated both OpenAPI contracts (type: string, format:
  decimal) so the contract stops silently disagreeing with the
  implementation - this was the actual root cause of the original gap.
- Test assertions updated to compare via Decimal instead of hardcoded
  literals; added 2 new regression tests for the near-miss precision case
  (10000.30 accepted, 10000.31 rejected) the old tests couldn't catch.

Sprint 6 event wiring (job.created, job.quoted, payment.completed):
- packages/events/__init__.py deleted. It violated STRUCTURE.md's own
  policy (packages/ is contracts-only) and would never have run in Docker
  - each service's Dockerfile only COPYs that service's own app/, never
  the repo-root packages/ folder. See ADR-0003.
- Each service gets its own app/events.py instead (lazy-connect,
  fire-and-forget Redis publish, unchanged design otherwise).
- Every publisher payload checked against its actual schema in
  packages/contracts/events/*.schema.json, not assumed from the deleted
  module's original function signatures. Fixed real gaps found doing
  this: job.created was missing required created_at, job.quoted was
  missing required quoted_at.
- job.completed intentionally NOT wired: its schema requires
  actual_depth_ft/actual_cost, and no endpoint captures that data yet
  (FR-TRACK-03 isn't built). Wire it up alongside that endpoint, not
  before - fabricating required fields would be worse than not firing it.
- job.status_changed dropped entirely: present in the old shared module
  but has no schema file and isn't one of the four events Development
  Plan Milestone 2 actually scopes. Undocumented event = scope creep,
  same as an undocumented endpoint would be (Development Plan section 3).
- Verified against a real Redis instance, not just "doesn't crash when
  Redis is absent": started redis-server, ran each service's real
  publisher, subscribed with a real Python pubsub client, and confirmed
  the wire payload matches each schema's required fields exactly.

Also: ci-frontend.yml's job matrix still listed the 3 deprecated apps
(contractor-app/customer-app/resource-owner-app) instead of web-app -
looks like an omission from the frontend-consolidation push, not
something from this change; fixed alongside since CI would otherwise
build directories that no longer exist.

Verified: 91 tests passing (89 + 2 new), ruff clean, contract-check clean,
across all 4 services.

+++ services/payments-data/app/schemas.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, ConfigDict, Field
 
@@ -7,7 +8,10 @@ from pydantic import BaseModel, ConfigDict, Field
 class PaymentCreateRequest(BaseModel):
     job_id: uuid.UUID
     quotation_id: uuid.UUID
-    amount: float = Field(gt=0)
+    # Decimal, not float: compared for exact equality against the quotation's
+    # total_estimate (SRS section 6). See fetch_quotation in
+    # app/payments/quotation_client.py for how the comparison side is parsed.
+    amount: Decimal = Field(gt=0)
     idempotency_key: str = Field(min_length=1)
 
 
@@ -15,7 +19,7 @@ class PaymentResponse(BaseModel):
     payment_id: uuid.UUID
     job_id: uuid.UUID
     quotation_id: uuid.UUID
-    amount: float
+    amount: Decimal
     status: str
     created_at: datetime
